stop and start stream in track struct, move track to track.rs
