+++ src-tauri/src/types.rs
@@ -1,9 +1,4 @@
-use std::{
-    collections::HashMap,
-    fs,
-    sync::{Arc, Mutex},
-};
-
+use std::collections::HashMap;
 use cpal::traits::{DeviceTrait, HostTrait};
 
 pub struct RingBuffer {
@@ -117,149 +112,3 @@ impl OutputDeviceRegistry {
 }
 
 // TODO add host manager
-
-pub trait TrackAudioSource: Send + Sync {
-    fn read(&mut self, buffer: &mut [f32]) -> usize;
-    fn write(&mut self, buffer: &[f32]) -> bool;
-}
-
-pub struct StreamSource {
-    ring_buffer: Arc<Mutex<RingBuffer>>,
-    stream: cpal::Stream,
-}
-
-impl StreamSource {
-    pub fn new(device: cpal::Device) -> Self {
-        let ring_buffer = Arc::new(Mutex::new(RingBuffer::new()));
-        let ring_buffer_clone = ring_buffer.clone();
-
-        if device.supports_input() {
-            let config = device
-                .default_input_config()
-                .expect("Failed to get input config");
-
-            let stream = device
-                .build_input_stream(
-                    &config.into(),
-                    move |data: &[f32], _: &cpal::InputCallbackInfo| {
-                        if let Ok(mut rb) = ring_buffer_clone.lock() {
-                            rb.write(data);
-                        }
-                    },
-                    move |err| eprintln!("Stream error: {}", err),
-                    None,
-                )
-                .expect("Failed to create input stream");
-
-            StreamSource {
-                ring_buffer,
-                stream,
-            }
-        } else if device.supports_output() {
-            let config = device
-                .default_output_config()
-                .expect("Failed to get output config");
-
-            let stream = device
-                .build_output_stream(
-                    &config.into(),
-                    move |data: &mut [f32], _: &cpal::OutputCallbackInfo| {
-                        for sample in data.iter_mut() {
-                            *sample = 0.0;
-                        }
-                    },
-                    move |err| eprintln!("Stream error: {}", err),
-                    None,
-                )
-                .expect("Failed to create output stream");
-
-            StreamSource {
-                ring_buffer,
-                stream,
-            }
-        } else {
-            panic!("")
-        }
-    }
-}
-
-pub struct FileSource {
-    reader: hound::WavReader<fs::File>,
-    writer: hound::WavWriter<fs::File>,
-}
-
-impl TrackAudioSource for StreamSource {
-    fn read(&mut self, buffer: &mut [f32]) -> usize {
-        if let Ok(mut rb) = self.ring_buffer.lock() {
-            rb.read(buffer)
-        } else {
-            0
-        }
-    }
-
-    fn write(&mut self, buffer: &[f32]) -> bool {
-        false
-    }
-}
-
-impl TrackAudioSource for FileSource {
-    fn read(&mut self, buffer: &mut [f32]) -> usize {
-        for (i, sample) in self.reader.samples::<f32>().take(buffer.len()).enumerate() {
-            buffer[i] = sample.unwrap_or(0.0);
-        }
-        buffer.len()
-    }
-
-    fn write(&mut self, buffer: &[f32]) -> bool {
-        false
-    }
-}
-
-pub enum TrackType {
-    MasterIn,
-    MasterOut,
-}
-
-pub struct Track {
-    pub track_type: TrackType,
-    pub source: Box<dyn TrackAudioSource + Send>,
-    pub volume: f32,
-    pub pan: f32,
-}
-
-impl Track {
-    pub fn new(track_type: TrackType, source: Box<dyn TrackAudioSource + Send>) -> Self {
-        Track {
-            track_type,
-            source,
-            volume: 100.0,
-            pan: 0.0,
-        }
-    }
-}
-
-pub struct TrackList {
-    tracks: HashMap<String, Arc<Mutex<Track>>>,
-}
-
-impl TrackList {
-    pub fn new() -> Self {
-        TrackList {
-            tracks: HashMap::new(),
-        }
-    }
-
-    pub fn add_track(&mut self, name: &str, track: Track) {
-        self.tracks.insert(name.into(), Arc::new(Mutex::new(track)));
-    }
-
-    pub fn remove_track(&mut self, name: &str) {
-        if self.tracks.contains_key(name) {
-            self.tracks.remove(name);
-        }
-    }
-
-    pub fn get_track(&self, name: &str) -> Option<&Arc<Mutex<Track>>> {
-        self.tracks.get(name)
-    }
-}
